test(server): cover run persistence + 404 to satisfy coverage gate
