+++ server/storage/runs.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import json
 import os
+import re
 import time
 import uuid
 from dataclasses import asdict, dataclass
@@ -9,6 +10,7 @@ from pathlib import Path
 from typing import Any, Dict, List, Optional
 
 RUNS_DIR = Path(os.getenv("GOLFIQ_RUNS_DIR", "data/runs")).resolve()
+RUN_ID_RE = r"^[0-9]{10}-[0-9a-f]{8}$"
 
 
 @dataclass(frozen=True)
@@ -24,6 +26,20 @@ class RunRecord:
 
 def _dir(run_id: str) -> Path:
     return RUNS_DIR / run_id
+
+
+def _safe(run_id: str) -> Optional[Path]:
+    if not re.fullmatch(RUN_ID_RE, run_id):
+        return None
+    resolved = (RUNS_DIR / run_id).resolve()
+    root = RUNS_DIR.resolve()
+    if not str(resolved).startswith(str(root)):
+        return None
+    try:
+        resolved.relative_to(root)
+    except ValueError:
+        return None
+    return resolved
 
 
 def save_run(
@@ -44,14 +60,20 @@ def save_run(
 
 
 def load_run(run_id: str) -> Optional[RunRecord]:
-    p = _dir(run_id) / "run.json"
+    safe_dir = _safe(run_id)
+    if not safe_dir:
+        return None
+    p = safe_dir / "run.json"
     if not p.exists():
         return None
     return RunRecord(**json.loads(p.read_text()))
 
 
 def delete_run(run_id: str) -> bool:
-    d = _dir(run_id)
+    safe_dir = _safe(run_id)
+    if not safe_dir:
+        return False
+    d = safe_dir
     if not d.exists():
         return False
     for c in d.iterdir():
